Criação do arquivo para novo produto e acesso a API
