+++ src/api/api.js
@@ -2,6 +2,7 @@ import axios from 'axios';
 
 const baseUrl = 'https://us-central1-labenu-apis.cloudfunctions.net/fourUsedThree/products';
 
+
 export const createProduct = async (body) => {
     try{
         return await axios.post(baseUrl,body)
